requirements.txt
fastapi==0.115.4
pydantic-settings==2.6.1
uvicorn==0.32.0
tenacity==9.0.0
loguru==0.7.2
opencv-python-headless==4.10.0.84
requests==2.32.3
pillow==11.0.0
google-genai==0.8.0